+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   * Updated react-router to v6
   * Bug fixes:
     * Fixed scroll when changing characters in party block (now page scrolls up)
+    * Go back button in the news block was hidden
 
 #### 1.1.6 February-2022
 * Changes